Add modularized security rules
