+++ gulpfile.js
@@ -52,7 +52,7 @@ var paths = {
 
   build: 'lib/',
 
-  curatedTypings: ['src/*.d.ts', '!src/instance-id.d.ts'],
+  curatedTypings: ['src/*.d.ts', '!src/instance-id.d.ts', '!src/security-rules.d.ts'],
 };
 
 const TEMPORARY_TYPING_EXCLUDES = [
@@ -67,7 +67,6 @@ const TEMPORARY_TYPING_EXCLUDES = [
   '!lib/messaging/*.d.ts',
   '!lib/project-management/*.d.ts',
   '!lib/remote-config/*.d.ts',
-  '!lib/security-rules/*.d.ts',
   '!lib/storage/*.d.ts',
   '!lib/utils/*.d.ts'
 ];
