+++ src/firebase-app.ts
@@ -34,6 +34,7 @@ import { InstanceIdImpl } from './instance-id/instance-id-internal';
 
 import { ProjectManagement } from './project-management/project-management';
 import { SecurityRules } from './security-rules/security-rules';
+import { SecurityRulesImpl } from './security-rules/security-rules-internal';
 import { RemoteConfig } from './remote-config/remote-config';
 
 import { Agent } from 'http';
@@ -391,8 +392,8 @@ export class FirebaseApp {
    */
   public securityRules(): SecurityRules {
     return this.ensureService_('security-rules', () => {
-      const securityRulesService: typeof SecurityRules =
-          require('./security-rules/security-rules').SecurityRules;
+      const securityRulesService: typeof SecurityRulesImpl =
+          require('./security-rules/security-rules-internal').SecurityRulesImpl;
       return new securityRulesService(this);
     });
   }
