+++ src/security-rules/security-rules-utils-internal.ts
