+++ src/security-rules/security-rules.ts
@@ -1,5 +1,5 @@
 /*!
- * Copyright 2019 Google Inc.
+ * Copyright 2020 Google Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,16 +14,13 @@
  * limitations under the License.
  */
 
-import { FirebaseServiceInterface, FirebaseServiceInternalsInterface } from '../firebase-service';
 import { FirebaseApp } from '../firebase-app';
-import * as validator from '../utils/validator';
-import {
-  SecurityRulesApiClient, RulesetResponse, RulesetContent, ListRulesetsResponse,
-} from './security-rules-api-client';
-import { FirebaseSecurityRulesError } from './security-rules-utils';
 
 /**
- * A source file containing some Firebase security rules.
+ * A source file containing some Firebase security rules. The content includes raw
+ * source code including text formatting, indentation and comments. Use the
+ * [`securityRules.createRulesFileFromSource()`](admin.securityRules.SecurityRules#createRulesFileFromSource)
+ * method to create new instances of this type.
  */
 export interface RulesFile {
   readonly name: string;
@@ -31,10 +28,19 @@ export interface RulesFile {
 }
 
 /**
- * Additional metadata associated with a Ruleset.
+ * Required metadata associated with a ruleset.
  */
 export interface RulesetMetadata {
+  /**
+   * Name of the `Ruleset` as a short string. This can be directly passed into APIs
+   * like [`securityRules.getRuleset()`](admin.securityRules.SecurityRules#getRuleset)
+   * and [`securityRules.deleteRuleset()`](admin.securityRules.SecurityRules#deleteRuleset).
+   */
   readonly name: string;
+
+  /**
+   * Creation time of the `Ruleset` as a UTC timestamp string.
+   */
   readonly createTime: string;
 }
 
@@ -46,319 +52,150 @@ export interface RulesetMetadataList {
   readonly nextPageToken?: string;
 }
 
-class RulesetMetadataListImpl implements RulesetMetadataList {
-
-  public readonly rulesets: RulesetMetadata[];
-  public readonly nextPageToken?: string;
-
-  constructor(response: ListRulesetsResponse) {
-    if (!validator.isNonNullObject(response) || !validator.isArray(response.rulesets)) {
-      throw new FirebaseSecurityRulesError(
-        'invalid-argument',
-        `Invalid ListRulesets response: ${JSON.stringify(response)}`);
-    }
-
-    this.rulesets = response.rulesets.map((rs) => {
-      return {
-        name: stripProjectIdPrefix(rs.name),
-        createTime: new Date(rs.createTime).toUTCString(),
-      };
-    });
-
-    if (response.nextPageToken) {
-      this.nextPageToken = response.nextPageToken;
-    }
-  }
+/**
+ * A set of Firebase security rules.
+ */
+export interface Ruleset extends RulesetMetadata {
+  readonly source: RulesFile[];
 }
 
 /**
- * Represents a set of Firebase security rules.
+ * The Firebase `SecurityRules` service interface.
+ *
+ * Do not call this constructor directly. Instead, use
+ * [`admin.securityRules()`](admin.securityRules#securityRules).
  */
-export class Ruleset implements RulesetMetadata {
-
-  public readonly name: string;
-  public readonly createTime: string;
-  public readonly source: RulesFile[];
-
-  constructor(ruleset: RulesetResponse) {
-    if (!validator.isNonNullObject(ruleset) ||
-      !validator.isNonEmptyString(ruleset.name) ||
-      !validator.isNonEmptyString(ruleset.createTime) ||
-      !validator.isNonNullObject(ruleset.source)) {
-      throw new FirebaseSecurityRulesError(
-        'invalid-argument',
-        `Invalid Ruleset response: ${JSON.stringify(ruleset)}`);
-    }
-
-    this.name = stripProjectIdPrefix(ruleset.name);
-    this.createTime = new Date(ruleset.createTime).toUTCString();
-    this.source = ruleset.source.files || [];
-  }
-}
-
-/**
- * SecurityRules service bound to the provided app.
- */
-export class SecurityRules implements FirebaseServiceInterface {
-
-  private static readonly CLOUD_FIRESTORE = 'cloud.firestore';
-  private static readonly FIREBASE_STORAGE = 'firebase.storage';
-
-  public readonly INTERNAL = new SecurityRulesInternals();
-
-  private readonly client: SecurityRulesApiClient;
+export interface SecurityRules {
+  app: FirebaseApp;
 
   /**
-   * @param {object} app The app for this SecurityRules service.
-   * @constructor
-   */
-  constructor(readonly app: FirebaseApp) {
-    this.client = new SecurityRulesApiClient(app);
-  }
-
-  /**
-   * Gets the Ruleset identified by the given name. The input name should be the short name string without
-   * the project ID prefix. For example, to retrieve the `projects/project-id/rulesets/my-ruleset`, pass the
-   * short name "my-ruleset". Rejects with a `not-found` error if the specified Ruleset cannot be found.
+   * Creates a {@link admin.securityRules.RulesFile `RuleFile`} with the given name
+   * and source. Throws an error if any of the arguments are invalid. This is a local
+   * operation, and does not involve any network API calls.
    *
-   * @param {string} name Name of the Ruleset to retrieve.
-   * @returns {Promise<Ruleset>} A promise that fulfills with the specified Ruleset.
+   * @example
+   * ```javascript
+   * const source = '// Some rules source';
+   * const rulesFile = admin.securityRules().createRulesFileFromSource(
+   *   'firestore.rules', source);
+   * ```
+   *
+   * @param name Name to assign to the rules file. This is usually a short file name that
+   *   helps identify the file in a ruleset.
+   * @param source Contents of the rules file.
+   * @return A new rules file instance.
    */
-  public getRuleset(name: string): Promise<Ruleset> {
-    return this.client.getRuleset(name)
-      .then((rulesetResponse) => {
-        return new Ruleset(rulesetResponse);
-      });
-  }
+  createRulesFileFromSource(name: string, source: string | Buffer): RulesFile;
 
   /**
-   * Gets the Ruleset currently applied to Cloud Firestore. Rejects with a `not-found` error if no Ruleset is
-   * applied on Firestore.
+   * Creates a new {@link admin.securityRules.Ruleset `Ruleset`} from the given
+   * {@link admin.securityRules.RulesFile `RuleFile`}.
    *
-   * @returns {Promise<Ruleset>} A promise that fulfills with the Firestore Ruleset.
+   * @param file Rules file to include in the new `Ruleset`.
+   * @returns A promise that fulfills with the newly created `Ruleset`.
    */
-  public getFirestoreRuleset(): Promise<Ruleset> {
-    return this.getRulesetForRelease(SecurityRules.CLOUD_FIRESTORE);
-  }
+  createRuleset(file: RulesFile): Promise<Ruleset>;
 
   /**
-   * Creates a new ruleset from the given source, and applies it to Cloud Firestore.
+   * Gets the {@link admin.securityRules.Ruleset `Ruleset`} identified by the given
+   * name. The input name should be the short name string without the project ID
+   * prefix. For example, to retrieve the `projects/project-id/rulesets/my-ruleset`,
+   * pass the short name "my-ruleset". Rejects with a `not-found` error if the
+   * specified `Ruleset` cannot be found.
    *
-   * @param {string|Buffer} source Rules source to apply.
-   * @returns {Promise<Ruleset>} A promise that fulfills when the ruleset is created and released.
+   * @param name Name of the `Ruleset` to retrieve.
+   * @return A promise that fulfills with the specified `Ruleset`.
    */
-  public releaseFirestoreRulesetFromSource(source: string | Buffer): Promise<Ruleset> {
-    return Promise.resolve()
-      .then(() => {
-        const rulesFile = this.createRulesFileFromSource('firestore.rules', source);
-        return this.createRuleset(rulesFile);
-      })
-      .then((ruleset) => {
-        return this.releaseFirestoreRuleset(ruleset)
-          .then(() => {
-            return ruleset;
-          });
-      });
-  }
+  getRuleset(name: string): Promise<Ruleset>;
 
   /**
-   * Makes the specified ruleset the currently applied ruleset for Cloud Firestore.
+   * Deletes the {@link admin.securityRules.Ruleset `Ruleset`} identified by the given
+   * name. The input name should be the short name string without the project ID
+   * prefix. For example, to delete the `projects/project-id/rulesets/my-ruleset`,
+   * pass the  short name "my-ruleset". Rejects with a `not-found` error if the
+   * specified `Ruleset` cannot be found.
    *
-   * @param {string|RulesetMetadata} ruleset Name of the ruleset to apply or a RulesetMetadata object containing
-   *   the name.
-   * @returns {Promise<void>} A promise that fulfills when the ruleset is released.
+   * @param name Name of the `Ruleset` to delete.
+   * @return A promise that fulfills when the `Ruleset` is deleted.
    */
-  public releaseFirestoreRuleset(ruleset: string | RulesetMetadata): Promise<void> {
-    return this.releaseRuleset(ruleset, SecurityRules.CLOUD_FIRESTORE);
-  }
+  deleteRuleset(name: string): Promise<void>;
 
   /**
-   * Gets the Ruleset currently applied to a Cloud Storage bucket. Rejects with a `not-found` error if no Ruleset is
-   * applied on the bucket.
+   * Retrieves a page of ruleset metadata.
    *
-   * @param {string=} bucket Optional name of the Cloud Storage bucket to be retrieved. If not specified,
-   *   retrieves the ruleset applied on the default bucket configured via `AppOptions`.
-   * @returns {Promise<Ruleset>} A promise that fulfills with the Cloud Storage Ruleset.
+   * @param pageSize The page size, 100 if undefined. This is also the maximum allowed
+   *   limit.
+   * @param nextPageToken The next page token. If not specified, returns rulesets
+   *   starting without any offset.
+   * @return A promise that fulfills with a page of rulesets.
    */
-  public getStorageRuleset(bucket?: string): Promise<Ruleset> {
-    return Promise.resolve()
-      .then(() => {
-        return this.getBucketName(bucket);
-      })
-      .then((bucketName) => {
-        return this.getRulesetForRelease(`${SecurityRules.FIREBASE_STORAGE}/${bucketName}`);
-      });
-  }
+  listRulesetMetadata(
+    pageSize?: number, nextPageToken?: string): Promise<RulesetMetadataList>;
 
   /**
-   * Creates a new ruleset from the given source, and applies it to a Cloud Storage bucket.
+   * Gets the {@link admin.securityRules.Ruleset `Ruleset`} currently applied to
+   * Cloud Firestore. Rejects with a `not-found` error if no ruleset is applied
+   * on Firestore.
    *
-   * @param {string|Buffer} source Rules source to apply.
-   * @param {string=} bucket Optional name of the Cloud Storage bucket to apply the rules on. If not specified,
-   *   applies the ruleset on the default bucket configured via `AppOptions`.
-   * @returns {Promise<Ruleset>} A promise that fulfills when the ruleset is created and released.
+   * @return A promise that fulfills with the Firestore ruleset.
    */
-  public releaseStorageRulesetFromSource(source: string | Buffer, bucket?: string): Promise<Ruleset> {
-    return Promise.resolve()
-      .then(() => {
-        // Bucket name is not required until the last step. But since there's a createRuleset step
-        // before then, make sure to run this check and fail early if the bucket name is invalid.
-        this.getBucketName(bucket);
-        const rulesFile = this.createRulesFileFromSource('storage.rules', source);
-        return this.createRuleset(rulesFile);
-      })
-      .then((ruleset) => {
-        return this.releaseStorageRuleset(ruleset, bucket)
-          .then(() => {
-            return ruleset;
-          });
-      });
-  }
+  getFirestoreRuleset(): Promise<Ruleset>;
 
   /**
-   * Makes the specified ruleset the currently applied ruleset for a Cloud Storage bucket.
+   * Creates a new {@link admin.securityRules.Ruleset `Ruleset`} from the given
+   * source, and applies it to Cloud Firestore.
    *
-   * @param {string|RulesetMetadata} ruleset Name of the ruleset to apply or a RulesetMetadata object containing
-   *   the name.
-   * @param {string=} bucket Optional name of the Cloud Storage bucket to apply the rules on. If not specified,
-   *   applies the ruleset on the default bucket configured via `AppOptions`.
-   * @returns {Promise<void>} A promise that fulfills when the ruleset is released.
+   * @param source Rules source to apply.
+   * @return A promise that fulfills when the ruleset is created and released.
    */
-  public releaseStorageRuleset(ruleset: string | RulesetMetadata, bucket?: string): Promise<void> {
-    return Promise.resolve()
-      .then(() => {
-        return this.getBucketName(bucket);
-      })
-      .then((bucketName) => {
-        return this.releaseRuleset(ruleset, `${SecurityRules.FIREBASE_STORAGE}/${bucketName}`);
-      });
-  }
+  releaseFirestoreRulesetFromSource(source: string | Buffer): Promise<Ruleset>;
 
   /**
-   * Creates a `RulesFile` with the given name and source. Throws if any of the arguments are invalid. This is a
-   * local operation, and does not involve any network API calls.
+   * Applies the specified {@link admin.securityRules.Ruleset `Ruleset`} ruleset
+   * to Cloud Firestore.
    *
-   * @param {string} name Name to assign to the rules file.
-   * @param {string|Buffer} source Contents of the rules file.
-   * @returns {RulesFile} A new rules file instance.
+   * @param ruleset Name of the ruleset to apply or a `RulesetMetadata` object
+   *   containing the name.
+   * @return A promise that fulfills when the ruleset is released.
    */
-  public createRulesFileFromSource(name: string, source: string | Buffer): RulesFile {
-    if (!validator.isNonEmptyString(name)) {
-      throw new FirebaseSecurityRulesError(
-        'invalid-argument', 'Name must be a non-empty string.');
-    }
-
-    let content: string;
-    if (validator.isNonEmptyString(source)) {
-      content = source;
-    } else if (validator.isBuffer(source)) {
-      content = source.toString('utf-8');
-    } else {
-      throw new FirebaseSecurityRulesError(
-        'invalid-argument', 'Source must be a non-empty string or a Buffer.');
-    }
-
-    return {
-      name,
-      content,
-    };
-  }
+  releaseFirestoreRuleset(ruleset: string | RulesetMetadata): Promise<void>;
 
   /**
-   * Creates a new `Ruleset` from the given `RulesFile`.
+   * Gets the {@link admin.securityRules.Ruleset `Ruleset`} currently applied to a
+   * Cloud Storage bucket. Rejects with a `not-found` error if no ruleset is applied
+   * on the bucket.
    *
-   * @param {RulesFile} file Rules file to include in the new Ruleset.
-   * @returns {Promise<Ruleset>} A promise that fulfills with the newly created Ruleset.
+   * @param bucket Optional name of the Cloud Storage bucket to be retrieved. If not
+   *   specified, retrieves the ruleset applied on the default bucket configured via
+   *   `AppOptions`.
+   * @return A promise that fulfills with the Cloud Storage ruleset.
    */
-  public createRuleset(file: RulesFile): Promise<Ruleset> {
-    const ruleset: RulesetContent = {
-      source: {
-        files: [ file ],
-      },
-    };
-
-    return this.client.createRuleset(ruleset)
-      .then((rulesetResponse) => {
-        return new Ruleset(rulesetResponse);
-      });
-  }
+  getStorageRuleset(bucket?: string): Promise<Ruleset>;
 
   /**
-   * Deletes the Ruleset identified by the given name. The input name should be the short name string without
-   * the project ID prefix. For example, to delete the `projects/project-id/rulesets/my-ruleset`, pass the
-   * short name "my-ruleset". Rejects with a `not-found` error if the specified Ruleset cannot be found.
+   * Creates a new {@link admin.securityRules.Ruleset `Ruleset`} from the given
+   * source, and applies it to a Cloud Storage bucket.
    *
-   * @param {string} name Name of the Ruleset to delete.
-   * @returns {Promise<Ruleset>} A promise that fulfills when the Ruleset is deleted.
+   * @param source Rules source to apply.
+   * @param bucket Optional name of the Cloud Storage bucket to apply the rules on. If
+   *   not specified, applies the ruleset on the default bucket configured via
+   *   {@link admin.AppOptions `AppOptions`}.
+   * @return A promise that fulfills when the ruleset is created and released.
    */
-  public deleteRuleset(name: string): Promise<void> {
-    return this.client.deleteRuleset(name);
-  }
+  releaseStorageRulesetFromSource(
+    source: string | Buffer, bucket?: string): Promise<Ruleset>;
 
   /**
-   * Retrieves a page of rulesets.
+   * Applies the specified {@link admin.securityRules.Ruleset `Ruleset`} ruleset
+   * to a Cloud Storage bucket.
    *
-   * @param {number=} pageSize The page size, 100 if undefined. This is also the maximum allowed limit.
-   * @param {string=} nextPageToken The next page token. If not specified, returns rulesets starting
-   *   without any offset.
-   * @returns {Promise<RulesetMetadataList>} A promise that fulfills a page of rulesets.
+   * @param ruleset Name of the ruleset to apply or a `RulesetMetadata` object
+   *   containing the name.
+   * @param bucket Optional name of the Cloud Storage bucket to apply the rules on. If
+   *   not specified, applies the ruleset on the default bucket configured via
+   *   {@link admin.AppOptions `AppOptions`}.
+   * @return A promise that fulfills when the ruleset is released.
    */
-  public listRulesetMetadata(pageSize = 100, nextPageToken?: string): Promise<RulesetMetadataList> {
-    return this.client.listRulesets(pageSize, nextPageToken)
-      .then((response) => {
-        return new RulesetMetadataListImpl(response);
-      });
-  }
-
-  private getRulesetForRelease(releaseName: string): Promise<Ruleset> {
-    return this.client.getRelease(releaseName)
-      .then((release) => {
-        const rulesetName = release.rulesetName;
-        if (!validator.isNonEmptyString(rulesetName)) {
-          throw new FirebaseSecurityRulesError(
-            'not-found', `Ruleset name not found for ${releaseName}.`);
-        }
-
-        return this.getRuleset(stripProjectIdPrefix(rulesetName));
-      });
-  }
-
-  private releaseRuleset(ruleset: string | RulesetMetadata, releaseName: string): Promise<void> {
-    if (!validator.isNonEmptyString(ruleset) &&
-      (!validator.isNonNullObject(ruleset) || !validator.isNonEmptyString(ruleset.name))) {
-      const err = new FirebaseSecurityRulesError(
-        'invalid-argument', 'ruleset must be a non-empty name or a RulesetMetadata object.');
-      return Promise.reject(err);
-    }
-
-    const rulesetName = validator.isString(ruleset) ? ruleset : ruleset.name;
-    return this.client.updateRelease(releaseName, rulesetName)
-      .then(() => {
-        return;
-      });
-  }
-
-  private getBucketName(bucket?: string): string {
-    const bucketName = (typeof bucket !== 'undefined') ? bucket :  this.app.options.storageBucket;
-    if (!validator.isNonEmptyString(bucketName)) {
-      throw new FirebaseSecurityRulesError(
-        'invalid-argument',
-        'Bucket name not specified or invalid. Specify a default bucket name via the ' +
-        'storageBucket option when initializing the app, or specify the bucket name ' +
-        'explicitly when calling the rules API.',
-      );
-    }
-
-    return bucketName;
-  }
-}
-
-class SecurityRulesInternals implements FirebaseServiceInternalsInterface {
-  public delete(): Promise<void> {
-    return Promise.resolve();
-  }
-}
-
-function stripProjectIdPrefix(name: string): string {
-  return name.split('/').pop()!;
+  releaseStorageRuleset(
+    ruleset: string | RulesetMetadata, bucket?: string): Promise<void>;
 }
